HAL_AVR: switch to maximum 256 byte buffers on UARTs
this makes the interrupt handler 8 bit, which saves enough cycles on
ArduCopter to be worthwhile.

We don't have any code that uses a buffer above 256 bytes.

+++ libraries/AP_HAL_AVR/UARTDriver.cpp
@@ -61,8 +61,8 @@ void AVRUARTDriver::begin(uint32_t baud, uint16_t rxSpace, uint16_t txSpace) {
 		if (0 == txSpace)
 			txSpace = _txBuffer->mask + 1;
 
-		if (rxSpace == (_rxBuffer->mask + 1) && 
-			txSpace == (_txBuffer->mask + 1)) {
+		if (rxSpace == (_rxBuffer->mask + 1U) && 
+                    txSpace == (_txBuffer->mask + 1U)) {
 			// avoid re-allocating the buffers if possible
 			need_allocate = false;
 			*_ucsrb &= ~(_portEnableBits | _portTxBits);
@@ -179,7 +179,7 @@ void AVRUARTDriver::flush(void) {
 }
 
 size_t AVRUARTDriver::write(uint8_t c) {
-	uint16_t i;
+	uint8_t i;
 
 	if (!_open) // drop bytes if not open
 		return 0;
@@ -212,7 +212,7 @@ size_t AVRUARTDriver::write(uint8_t c) {
 
 bool AVRUARTDriver::_allocBuffer(Buffer *buffer, uint16_t size)
 {
-	uint16_t mask;
+	uint8_t mask;
 	uint8_t	 shift;
 
 	// init buffer state
